+++ packages/substrate/src/wanxiang_substrate/lineage/__init__.py
@@ -1,0 +1,9 @@
+"""World Lineage substrate (M28)."""
+
+from wanxiang_substrate.lineage.graph import (
+    LineageEdge,
+    LineageGraph,
+    LineageNode,
+)
+
+__all__ = ["LineageEdge", "LineageGraph", "LineageNode"]